fix(test): 단위 테스트 DB 연결 차단 + 확장 ingest 서비스 분할 (CI Gate 3)
**테스트 플레이키** — 회차마다 무관한 스위트(approval-gate·rate-limiter-behavior·
spawn-agents 등)가 1건씩 랜덤 실패하다 단독 실행하면 통과하던 문제.

원인: jest.setup.ts 가 DATABASE_URL 을 지워 "테스트는 운영 DB 를 건드리지 않는다"를
의도했지만, config/env.ts 의 기본값(postgresql://localhost:5432/openmake_llm)이 그
자리를 메워 **실제로 로컬 DB 에 접속**하고 있었다. 그 연결이 테스트 종료 후까지 살아남아
teardown 뒤 pg 가 pgPass 를 lazy require 하며 던지는 unhandled error 가 그때 실행 중이던
**무관한 스위트에 귀속**돼 실패로 집계됐다.

수정: setupFilesAfterEnv 에서 TEST_DATABASE_URL 이 없으면 pg 의 Pool/Client 를 거부
스텁으로 대체해 소켓을 아예 열지 않는다. DB 통합 테스트는 TEST_DATABASE_URL 을 주면
실제 pg 를 그대로 쓴다(없으면 기존 관용구대로 describe.skip). 5회 연속 FAIL 0 ·
teardown 누수 0 확인, 실행 시간도 47s → 42s.

**CI Gate 3(600줄) 위반 해소** — Phase 2 로 extension-ingest-service.ts 가 829줄이 되어
CI 파일 크기 가드를 깨뜨리는 상태였다(eslint 400 경고와 별개인 CI 전용 게이트).
구성요소 설치 네 갈래(commands→스킬 / agents→Custom Agent / 스킬 번들 파일 / MCP draft)를
extension-components.ts 로, 공개 타입을 extension-ingest-types.ts 로 분리:
829 → 488줄(+ 379 + 97). 동작 변경 없음 — 기존 import 경로는 타입 재노출로 호환.
